refactor: extract network helpers into network service
